Replace XcmError::Undefined usage.
