+++ xcm-support/src/currency_adapter.rs
@@ -1,0 +1,80 @@
+use codec::FullCodec;
+use sp_runtime::traits::{MaybeSerializeDeserialize, SaturatedConversion};
+use sp_std::{
+	cmp::{Eq, PartialEq},
+	fmt::Debug,
+	marker::PhantomData,
+	prelude::*,
+	result,
+};
+
+use xcm::v0::{Error as XcmError, MultiAsset, MultiLocation, Result};
+use xcm_executor::traits::{LocationConversion, MatchesFungible, TransactAsset};
+
+use crate::CurrencyIdConversion;
+
+/// Asset transaction errors.
+enum Error {
+	/// Asset not found.
+	AssetNotFound,
+	/// `MultiLocation` to `AccountId` Conversion failed.
+	AccountIdConversionFailed,
+	/// `CurrencyId` conversion failed.
+	CurrencyIdConversionFailed,
+}
+
+impl From<Error> for XcmError {
+	fn from(e: Error) -> Self {
+		match e {
+			Error::AssetNotFound => XcmError::FailedToTransactAsset("AssetNotFound"),
+			Error::AccountIdConversionFailed => XcmError::FailedToTransactAsset("AccountIdConversionFailed"),
+			Error::CurrencyIdConversionFailed => XcmError::FailedToTransactAsset("CurrencyIdConversionFailed"),
+		}
+	}
+}
+
+pub struct MultiCurrencyAdapter<MultiCurrency, Matcher, AccountIdConverter, AccountId, CurrencyIdConverter, CurrencyId>(
+	PhantomData<(
+		MultiCurrency,
+		Matcher,
+		AccountIdConverter,
+		AccountId,
+		CurrencyIdConverter,
+		CurrencyId,
+	)>,
+);
+
+impl<
+		MultiCurrency: orml_traits::MultiCurrency<AccountId, CurrencyId = CurrencyId>,
+		Matcher: MatchesFungible<MultiCurrency::Balance>,
+		AccountIdConverter: LocationConversion<AccountId>,
+		AccountId: sp_std::fmt::Debug,
+		CurrencyIdConverter: CurrencyIdConversion<CurrencyId>,
+		CurrencyId: FullCodec + Eq + PartialEq + Copy + MaybeSerializeDeserialize + Debug,
+	> TransactAsset
+	for MultiCurrencyAdapter<MultiCurrency, Matcher, AccountIdConverter, AccountId, CurrencyIdConverter, CurrencyId>
+{
+	fn deposit_asset(asset: &MultiAsset, location: &MultiLocation) -> Result {
+		let who =
+			AccountIdConverter::from_location(location).ok_or(XcmError::from(Error::AccountIdConversionFailed))?;
+		let currency_id =
+			CurrencyIdConverter::from_asset(asset).ok_or(XcmError::from(Error::CurrencyIdConversionFailed))?;
+		let amount: MultiCurrency::Balance = Matcher::matches_fungible(&asset)
+			.ok_or(XcmError::from(Error::AssetNotFound))?
+			.saturated_into();
+		MultiCurrency::deposit(currency_id, &who, amount).map_err(|e| XcmError::FailedToTransactAsset(e.into()))?;
+		Ok(())
+	}
+
+	fn withdraw_asset(asset: &MultiAsset, location: &MultiLocation) -> result::Result<MultiAsset, XcmError> {
+		let who =
+			AccountIdConverter::from_location(location).ok_or(XcmError::from(Error::AccountIdConversionFailed))?;
+		let currency_id =
+			CurrencyIdConverter::from_asset(asset).ok_or(XcmError::from(Error::CurrencyIdConversionFailed))?;
+		let amount: MultiCurrency::Balance = Matcher::matches_fungible(&asset)
+			.ok_or(XcmError::from(Error::AssetNotFound))?
+			.saturated_into();
+		MultiCurrency::withdraw(currency_id, &who, amount).map_err(|e| XcmError::FailedToTransactAsset(e.into()))?;
+		Ok(asset.clone())
+	}
+}
